把选择方块的逻辑从Switch切换为使用Resources查找字符串来选择

+++ Assets/Scripts/Manager/CameraHandler.cs
@@ -69,7 +69,7 @@ public class CameraHandler : MonoBehaviour
 
     }
 
-    public void ZoomInToShip(int amount)
+    public void ZoomInToShip(float amount)
     {
         targetOrthographicSize = amount;
     }
